Add config directory and moved most of index.js to it in order to increase readability and maintainability
